+++ R/read_sigaa.R
@@ -17,9 +17,6 @@
 #' Be sure that your data has the variables: "Matricula", "Nome", "Situacao Matricula", 
 #' "Curso", "Cpf", "Instituicao", "Per. Letivo Inicial".
 #' 
-#' @examples  
-#' # this dataset is not a real one. It is just for test purpose.
-#' sigaa
 #' @importFrom dplyr %>% sym 
 #' @export
 read_sigaa <- function(path = ""){