+++ tests/testthat/test-read_files.R
@@ -20,4 +20,3 @@ test_that("read_qacademico works", {
   check_rfept_table(qacademico, expect_nrow = 14366)
   expect_true(inherits(qacademico, "rfept_data_frame"))
 })
-